+++ lib/parse/client.rb
@@ -189,13 +189,13 @@ module Parse
       # http method
       method = method.downcase.to_sym
       # set the User-Agent
-      headers["User-Agent".freeze] = "Parse-Stack Ruby Client v#{Parse::Stack::VERSION}".freeze
+      headers["User-Agent".freeze] = "Parse-Server Ruby Client v#{Parse::Stack::VERSION}".freeze
 
       if opts[:cache] == false
-        headers[Parse::Middleware::Caching::CACHE_CONTROL] = "no-cache"
+        headers[Parse::Middleware::Caching::CACHE_CONTROL] = "no-cache".freeze
       elsif opts[:cache].is_a?(Numeric)
-        # future feature
-        # headers[Parse::Middleware::Caching::CACHE_CONTROL] = "max-age: #{opts[:cache].to_i}"
+        # specify the cache duration of this request
+        headers[Parse::Middleware::Caching::CACHE_EXPIRES_DURATION] = opts[:cache].to_i
       end
 
       if opts[:use_master_key] == false